requirements.txt
pywin32
boto3
pymssql<3.0
filechunkio
